add test and refactor id type

+++ platform/src/main/java/com/eventplatform/factory/UserFactory.java
@@ -2,7 +2,6 @@ package com.eventplatform.factory;
 
 import com.eventplatform.model.User;
 
-import java.math.BigInteger;
 import java.util.Date;
 
 public class UserFactory {
@@ -11,7 +10,7 @@ public class UserFactory {
         return new User(user.getId(), user.getCreationDate(), user.getName(), user.getSurname(), user.getLogin(), user.getEmail(), user.getPassword());
     }
 
-    public static User createUser(BigInteger id, String name, String surname, String login, String email, String password) {
+    public static User createUser(int id, String name, String surname, String login, String email, String password) {
         return new User(id, new Date(System.currentTimeMillis()), name, surname, login, email, password);
     }
 }
